flask_app/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Pillow==10.2.0
torch==2.3.0
torchvision==0.18.0
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.1
google-cloud-storage==2.16.0